adds clockskew class

+++ SpeakerSocial/Audio.h
@@ -3,7 +3,7 @@
 @property (strong, nonatomic) AVAudioPlayer *audioPlayer;
 - (void)play:(double)position;
 - (void)load;
-- (NSInteger)duration;
+- (double)duration;
 - (NSTimeInterval)currentTime;
 - (void)mute;
 @end
